Fixed bugs with photo upload and photo delete. Photo upload still not saving input of name and url, however.
