+++ src/components/Photos.js
@@ -75,7 +75,7 @@ const Photos = () => {
 
   return (
     <div>
-      <span class='albumHeader'>
+      <span className='albumHeader'>
       <FontAwesomeIcon 
         icon={faArrowLeft} 
         className='logos' 
@@ -97,6 +97,7 @@ const Photos = () => {
         show={modalShow}
         onHide={() => setModalShow(false)}
         albumId={albumId}
+        albumName={albumName}
         />
           {/* Trash icon, open edit modal */}
         </span>
